backend/requirements.txt
Flask
python-dotenv
flask-cors
requests
g4f
beautifulsoup4